blub

+++ gameWindow.py
@@ -175,7 +175,7 @@ class DraughtsGameWindow(QWidget):
 				
 				transformer = tf.ProjectiveTransform()
 				transformer.estimate(src, dst)
-				transformed_image = tf.warp(input_img, transformer, output_shape=(800, 800))
+				transformed_image = tf.warp(input_img, transformer, output_shape=(800, 800)).astype(int)
 				Image.fromarray(transformed_image).save('./transformed.png')
 				
 		camera.close()
